pyproject.toml
[build-system]
requires = ["uv_build>=0.7.12,<0.8.0"]
build-backend = "uv_build"

[project]
name = "hyundai_wia"
version = "0.0.1"
authors = [{ name = "sunghoon.kim", email = "sunghoon014@gmail.com" }]
packages = [{ include = "app" }]
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "ruff>=0.5.0,<0.6.0",
    "loguru>=0.7.3,<0.8.0",
    "fastapi>=0.115.0",
    "dependency-injector>=4.41.0,<5.0.0",
    "pydantic-settings>=2.3.4,<3.0.0",
    "openai>=1.35.10,<2.0.0",
    "langchain-openai>=0.3.30",
    "langchain-experimental>=0.3.4",
    "langgraph>=0.5.3",
    "langfuse==2.60.9",
    "numpy>=1.26.0,<2.0.0",
    "tiktoken>=0.7.0,<0.8.0",
    "pandas>=2.2.2,<3.0.0",
    "tqdm>=4.66.4,<5.0.0",
    "aiohttp>=3.10.3,<4.0.0",
    "fastmcp>=2.9.2",
    "docker>=7.1.0",
    "pymilvus>=2.5.12",
    "nltk>=3.9.1",
    "motor>=3.7.1",
    "pymupdf>=1.26.3",
    "ipykernel>=6.30.1",
    "requests>=2.32.4",
    "paho-mqtt>=2.1.0",
]

[tool.uv]
package = false

[tool.ruff.lint]
# 활성화할 규칙들
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "D",   # pydocstyle
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

# 무시할 규칙들
ignore = [
    "D100", # Missing docstring in public module
    "D101", # Missing docstring in public class
    "D102", # Missing docstring in public method
    "D103", # Missing docstring in public function
    "D104", # Missing docstring in public package
    "D105", # Missing docstring in magic method
    "D106", # Missing docstring in public nested class
    "D107", # Missing docstring in __init__
    "E501", # Line too long (handled by formatter)
    "B008", # Do not perform function calls in argument defaults
]

# 자동 수정하지 않을 규칙들
unfixable = ["F401", "F841"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["app", "ml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers --import-mode=importlib -s -v"
markers = ["asyncio", "parametrize"]
